Database file for alx bookstore

+++ alx_book_store.sql
@@ -32,7 +32,7 @@ CREATE TABLE Orders_Details (
 order_detail_id INT AUTO_INCREMENT PRIMARY KEY,
 order_id INT,
 book_id INT,
-quantity INT,
-FOREIGN KEY (order_id) REFERENCES Orders (order_id),
+quantity Double (10,2),
+FOREIGN KEY (order_id) REFERENCES Orders(order_id),
 FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
